Maintenance events removed, only critical situation events considered due to lack of time and relevance to overall project goal
