+++ src/Event/Event.h
@@ -3,6 +3,10 @@
 
 #include "../Time/Time.h"
 #include <cstdint>
+#include "Event.h"
+#include "../Location/Location.h"
+class Vehicle;
+#include <vector>
 
 namespace LockFreeDispatch {
 
@@ -12,6 +16,9 @@ namespace LockFreeDispatch {
         Time startTime;
         uint64_t durationSeconds;
         bool dispatched;
+        Location eventLocation;
+        std::vector<Vehicle*> vehicleRequirements;
+        uint16_t vehicleRequirementId;
     public:
         Event();
 
@@ -30,6 +37,20 @@ namespace LockFreeDispatch {
         bool getDispatched() const;
 
         void setDispatched(bool dispatchStatus);
+
+        Location getLocation();
+
+        void setLocation(Location newLocation);
+
+        std::vector<Vehicle*> getVehicleRequirements();
+
+        void addVehicleRequirementId(uint16_t newVehicleRequirementId);
+
+        Vehicle *popVehicleRequirement();
+
+        void deleteVehicleRequirements();
+
+        bool findVehicleRequirement(Vehicle *vehicle);
     };
 
 } // LockFreeDispatch
